backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
pytineye==3.0.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
slowapi==0.1.9
